feat: Expert User Info
