+++ backend-server/.gitignore
@@ -34,6 +34,7 @@ application-debug.yml
 application-prod.yml
 *.p8
 *.p12
+*.jks
 
 ### NetBeans ###
 /nbproject/private/
