+++ frontend-web/src/api/auth/experts/types/expert-user.ts
@@ -1,10 +1,12 @@
 import { EJob } from "@type/job.ts";
 
 export type TExpertUser = {
-  name: string;
-  birthday: number[];
-  tel: string;
+  name: string | null;
+  birthday: number[] | null;
+  tel: string | null;
   job: EJob | null;
   department: string | null;
   belong: string | null;
+  isOptionalAgreementAccepted: boolean | null;
+  isIdentified: boolean;
 };
