+++ frontend-web/src/components/login/after-social-login/view.tsx
@@ -1,0 +1,9 @@
+import LoadingSpinner from "@/components/loading-spinner/view";
+import { useAfterSocialLoginViewController } from "@components/login/after-social-login/view.controller.ts";
+
+function AfterSocialLoginPage() {
+  useAfterSocialLoginViewController();
+  return <LoadingSpinner />;
+}
+
+export default AfterSocialLoginPage;
