+++ frontend-web/src/router.tsx
@@ -12,7 +12,7 @@ import PatientPage from "@components/main/patient/view.tsx";
 import SocialLoginFailurePage from "@components/login/social-login-failure/view.tsx";
 import { RegistrationPage as ExpertFacilityRegistration } from "@components/main/registration/view.tsx";
 import MyPage from "@components/main/mypage/view.tsx";
-import LoginProcessJunctionPage from "@components/login/login-process-junction/view.tsx";
+import AfterSocialLoginPage from "@components/login/after-social-login/view.tsx";
 import LoadingPage from "@page/loading/view.tsx";
 import IdentifyPage from "@components/login/identify/view.tsx";
 import IdentifyResultPage from "@components/login/identify-result/view.tsx";
@@ -102,6 +102,14 @@ export const router = createBrowserRouter([
         ],
       },
       {
+        path: "login/social/kakao",
+        element: <AfterSocialLoginPage />,
+      },
+      {
+        path: "login/social/google",
+        element: <AfterSocialLoginPage />,
+      },
+      {
         path: "",
         element: <ExpertLogin />,
         errorElement: <NotFoundPage />,
@@ -124,14 +132,6 @@ export const router = createBrowserRouter([
               {
                 path: "login/social/not-yet",
                 element: <SocialLoginNotYetPage />,
-              },
-              {
-                path: "login/social/kakao",
-                element: <LoginProcessJunctionPage />,
-              },
-              {
-                path: "login/social/google",
-                element: <LoginProcessJunctionPage />,
               },
               {
                 path: "login/identify",
